Income rise and growth of prison population
Exploration of the relationship between the growth of the median income and America's incardination rate. Here is was determined that if the growth rate continued at the same rate as it did in the 1980's there would 6504 persons per 100,000 in prison when the median reaches $100,000. Skills demonstrated here include advanced visualizations and the use of correlation and prediction functions.